+++ apps/web/components/Feature/ProfileLayout/HomeSections/styles.ts
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+import {
+  ContentInner,
+  HeroWrapper,
+} from "@/components/Feature/ProfileLayout/Hero/styles";
+
+export const SectionWrapper = styled(HeroWrapper)`
+  margin-top: 0;
+  padding-top: 35px;
+  background: transparent;
+`;
+
+export const ContentAdjust = styled(ContentInner)`
+  padding: 0 6px;
+`;
