+++ apps/web/components/Feature/ProfileLayout/Shell.tsx
@@ -1,0 +1,22 @@
+"use client";
+
+import type { ReactNode } from "react";
+import Footer from "@/components/Feature/ProfileLayout/shared/Footer";
+import ProfileNavbar from "@/components/Feature/ProfileLayout/Navbar";
+import type { ProfileLayoutVariant } from "@/components/Feature/ProfileLayout/config";
+import { Page as PageShell } from "@/components/Feature/ProfileLayout/pageStyles";
+
+type ProfileShellProps = {
+  variant: ProfileLayoutVariant;
+  children: ReactNode;
+};
+
+export default function ProfileShell({ variant, children }: ProfileShellProps) {
+  return (
+    <PageShell>
+      <ProfileNavbar variant={variant} />
+      {children}
+      <Footer />
+    </PageShell>
+  );
+}
