--- apps/web/components/Layout/CreateProfile2/index.tsx
@@ -1,22 +1,0 @@
-"use client";
-
-import type { ReactNode } from "react";
-import { Page } from "../../Feature/ProfileLayout/Layout2/Home/styles";
-import Footer from "@/components/Feature/ProfileLayout/Layout1/Footer";
-import Navbar from "@/components/Feature/ProfileLayout/Layout2/Navbar";
-
-type CreateProfileLayoutProps = {
-  children: ReactNode;
-};
-
-export default function CreateProfileLayout({
-  children,
-}: CreateProfileLayoutProps) {
-  return (
-    <Page>
-      <Navbar />
-      {children}
-      <Footer />
-    </Page>
-  );
-}
